# intern strings extracted from small dictionaries