`StringDictionary.getStringValue` shows up in most profiles I have seen of Pinot. It used to be bottlenecked on finding the null terminator in `FixedByteValueReaderWriter.getUnpaddedString` but this was sped up in #7708, now the number of allocations is the bottleneck.

This PR adds an interning table in `StringDictionary` which is only used when the size of the strings (excluding object headers and shallow layout) to be interned is guaranteed to be less than 10MB, which is roughly equivalent to the block sizes used by `TransformFunction`s. When the dictionary fits into the intern table, this improves throughput (over 10x in some cases) eliminates allocation and GC time. When the dictionary is too large to be interned within 10MB, throughput and allocation rate do not regress.

Without interning table
```
Benchmark                                                                                   (_length)  (_nativeOrder)  (_paddingByte)  (_values)  Mode  Cnt           Score           Error   Units
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                       8            true              42        512  avgt    5          14.608 ±         0.112   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                   8            true              42        512  avgt    5       36304.396 ±         2.842    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                              8            true              42        512  avgt    5          28.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                       8            true              42       4096  avgt    5         139.851 ±         0.646   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                   8            true              42       4096  avgt    5      290683.783 ±        27.218    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                              8            true              42       4096  avgt    5          25.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                       8            true              42       8192  avgt    5         287.664 ±         9.572   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                   8            true              42       8192  avgt    5      581535.774 ±        55.799    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                              8            true              42       8192  avgt    5          24.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                       8            true              42      16384  avgt    5         566.878 ±        14.509   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                   8            true              42      16384  avgt    5     1162791.378 ±       110.594    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                              8            true              42      16384  avgt    5          24.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                       8           false              42        512  avgt    5          13.909 ±         0.202   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                   8           false              42        512  avgt    5       36304.378 ±         2.709    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                              8           false              42        512  avgt    5          30.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                       8           false              42       4096  avgt    5         141.954 ±         4.906   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                   8           false              42       4096  avgt    5      290683.864 ±        27.699    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                              8           false              42       4096  avgt    5          23.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                       8           false              42       8192  avgt    5         282.691 ±        12.511   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                   8           false              42       8192  avgt    5      581535.614 ±        54.560    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                              8           false              42       8192  avgt    5          23.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                       8           false              42      16384  avgt    5         565.192 ±        20.916   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                   8           false              42      16384  avgt    5     1162791.245 ±       109.161    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                              8           false              42      16384  avgt    5          25.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                      32            true              42        512  avgt    5          21.390 ±         0.539   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                  32            true              42        512  avgt    5       54672.574 ±         4.117    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                             32            true              42        512  avgt    5          28.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                      32            true              42       4096  avgt    5         209.035 ±        13.126   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                  32            true              42       4096  avgt    5      438077.624 ±        40.415    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                             32            true              42       4096  avgt    5          24.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                      32            true              42       8192  avgt    5         400.035 ±        12.125   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                  32            true              42       8192  avgt    5      875754.949 ±        78.816    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                             32            true              42       8192  avgt    5          25.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                      32            true              42      16384  avgt    5         818.458 ±        12.627   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                  32            true              42      16384  avgt    5     1754046.124 ±       158.981    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                             32            true              42      16384  avgt    5          24.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                      32           false              42        512  avgt    5          22.899 ±         0.597   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                  32           false              42        512  avgt    5       54672.625 ±         4.468    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                             32           false              42        512  avgt    5          26.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                      32           false              42       4096  avgt    5         204.732 ±         8.169   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                  32           false              42       4096  avgt    5      438077.533 ±        39.776    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                             32           false              42       4096  avgt    5          25.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                      32           false              42       8192  avgt    5         402.713 ±        10.173   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                  32           false              42       8192  avgt    5      875754.719 ±        78.094    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                             32           false              42       8192  avgt    5          26.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                      32           false              42      16384  avgt    5         897.035 ±         9.160   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                  32           false              42      16384  avgt    5     1754047.764 ±       174.105    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                             32           false              42      16384  avgt    5          24.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                    1024            true              42        512  avgt    5         142.570 ±        10.162   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                1024            true              42        512  avgt    5      808595.898 ±        28.537    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                           1024            true              42        512  avgt    5          36.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                    1024            true              42       4096  avgt    5        1184.816 ±        29.738   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                1024            true              42       4096  avgt    5     6511319.379 ±       229.643    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                           1024            true              42       4096  avgt    5          54.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                    1024            true              42       8192  avgt    5        2500.836 ±        95.902   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                1024            true              42       8192  avgt    5    13036130.134 ±       482.732    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                           1024            true              42       8192  avgt    5          52.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                    1024            true              42      16384  avgt    5        5789.566 ±       450.470   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                1024            true              42      16384  avgt    5    26110782.836 ±      1100.930    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                           1024            true              42      16384  avgt    5          38.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                    1024           false              42        512  avgt    5         149.560 ±         2.794   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                1024           false              42        512  avgt    5      808595.962 ±        28.838    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                           1024           false              42        512  avgt    5          54.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                    1024           false              42       4096  avgt    5        1248.535 ±        38.234   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                1024           false              42       4096  avgt    5     6511321.162 ±       242.151    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                           1024           false              42       4096  avgt    5          52.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                    1024           false              42       8192  avgt    5        2657.626 ±       221.590   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                1024           false              42       8192  avgt    5    13036133.971 ±       511.855    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                           1024           false              42       8192  avgt    5          48.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                    1024           false              42      16384  avgt    5        5959.743 ±       544.894   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm                1024           false              42      16384  avgt    5    26110794.161 ±      1192.076    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                           1024           false              42      16384  avgt    5          38.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                   65536            true              42        512  avgt    5       10276.808 ±       116.498   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm               65536            true              42        512  avgt    5    50814880.376 ±      1994.851    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                          65536            true              42        512  avgt    5          51.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                   65536            true              42       4096  avgt    5       81733.900 ±       199.826   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm               65536            true              42       4096  avgt    5   405690051.938 ±     15044.988    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                          65536            true              42       4096  avgt    5          53.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                   65536            true              42       8192  avgt    5      163259.840 ±      9478.884   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm               65536            true              42       8192  avgt    5   809156906.286 ±     27907.233    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                          65536            true              42       8192  avgt    5          55.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                   65536            true              42      16384  avgt    5      321912.650 ±      4214.615   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm               65536            true              42      16384  avgt    5  1615953222.800 ±     48827.319    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                          65536            true              42      16384  avgt    5          65.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                   65536           false              42        512  avgt    5       10318.616 ±       163.180   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm               65536           false              42        512  avgt    5    50790311.946 ±      2014.396    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                          65536           false              42        512  avgt    5          50.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                   65536           false              42       4096  avgt    5       81945.555 ±      1225.584   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm               65536           false              42       4096  avgt    5   405493485.785 ±     15044.999    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                          65536           false              42       4096  avgt    5          54.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                   65536           false              42       8192  avgt    5      163695.100 ±     10944.886   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm               65536           false              42       8192  avgt    5   808764318.743 ±     32678.816    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                          65536           false              42       8192  avgt    5          56.000                      ms
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary                                   65536           false              42      16384  avgt    5      322787.764 ±      4244.671   us/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.alloc.rate.norm               65536           false              42      16384  avgt    5  1615166949.200 ±     48882.435    B/op
BenchmarkFixedByteValueReaderWriter.readStringsFromDictionary:·gc.time                          65536           false              42      16384  avgt    5          65.000                      ms
```

With interning table
```
…