@richardstartin 2022-02-01
> High level question: the improvement in this PR is very similar to the `OnHeapStringDictionary`. I am thinking maybe we should just use `OnHeapStringDictionary` if the dictionary size is small

I think it’s worth comparison. This has a couple of advantages:
- fewer types, making it more likely the calls to getStringValue inline into hot loops, if they don’t inline it affects what can be done with that loop.
- Only store the values which actually get unpacked