@richardstartin 2022-02-01
Looking at `OnHeapStringDictionary`, it's much heavier than an intern table:

```java
  private final String[] _unpaddedStrings;
  private final Object2IntOpenHashMap<String> _unPaddedStringToIdMap;
  private final String[] _paddedStrings;
```

The padded strings and string to id map don't serve much purpose on the hot path.