@Jackie-Jiang 2022-02-01
The `_paddedStrings` should not be setup in normal case (it's for backward-compatibility for some very old segment format). The map is used for looking up dict id, which is used for filtering phase (less frequently called comparing to read value from dict id).
After a second thought, I feel adding this on-heap array to the default dictionary can potentially cause problems. Say a server has 10K segments loaded, each segment has one 10M string dictionary, it could consume 100G heap memory in the worst case