@richardstartin 2022-02-01
Yes, and many of the strings will be duplicated across many of those 10K segments too. You illustrate well how much memory bandwidth the query layer requires, because each transform function in each query for each segment will construct a similarly sized array. I think it would be better to introduce a `StringView` type which refers to the bytes without needing to copy them.